[REFACTOR] 내 점수와 전체 점수표로 섹션 분리
